backend/app/schemas/document_type.py
from typing import Optional
from pydantic import BaseModel, ConfigDict


class DocumentTypeCreate(BaseModel):
    name: str
    description: Optional[str] = None


class DocumentTypeOut(BaseModel):
    id: int
    name: str
    description: Optional[str] = None
    is_active: bool

    model_config = ConfigDict(from_attributes=True)