backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlmodel==0.0.24
psycopg[binary]==3.2.6
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
pydantic-settings==2.8.1
email-validator==2.2.0